Hmmm...

+++ src/main/java/cc/slack/features/modules/impl/combat/Velocity.java
@@ -114,7 +114,7 @@ public class Velocity extends Module {
                 break;
             case "hypixel damage strafe":
                 if (mc.getPlayer().hurtTime == 9) {
-                    MovementUtil.strafe((float) MovementUtil.getSpeed() * 0.75f);
+                    MovementUtil.strafe(MovementUtil.getSpeed() * 0.75f);
                 }
                 break;
             case "hypixel":
